Move speaker type to type folder
